lottie/vector: remove a warning message from 64 bits compilation
uint64_t is same as unsigned long long in 32 bits compilation.
But, it is same as unsigned long in 64 bits compilation.

+++ src/vector/vdebug.cpp
@@ -32,7 +32,7 @@ void format_timestamp(std::ostream& os, uint64_t timestamp)
     char        buffer[32];
     strftime(buffer, 32, "%Y-%m-%d %T.", gmtime);
     char microseconds[7];
-    sprintf(microseconds, "%06llu", timestamp % 1000000);
+    sprintf(microseconds, "%06llu", (long long unsigned int)timestamp % 1000000);
     os << '[' << buffer << microseconds << ']';
 }
